move pathfinding lib stuffs to their own folder in lib
