+++ src/main/java/org/sciborgs1155/robot/commands/Alignment.java
@@ -17,8 +17,8 @@ import java.util.function.Supplier;
 import org.sciborgs1155.lib.FaultLogger;
 import org.sciborgs1155.lib.FaultLogger.Fault;
 import org.sciborgs1155.lib.FaultLogger.FaultType;
-import org.sciborgs1155.lib.RepulsorFieldPlanner;
 import org.sciborgs1155.lib.Tracer;
+import org.sciborgs1155.lib.pathfinding.RepulsorFieldPlanner;
 import org.sciborgs1155.robot.drive.Drive;
 import org.sciborgs1155.robot.drive.DriveConstants;
 import org.sciborgs1155.robot.drive.DriveConstants.Translation;
